Change the trace storage so that the process-level trace has a unique extension.
We need to do this because the file open dialogue on Mac OS X only allows patterns on the extension of the file (so we can't use ".p.trace").
Also change the trace storage so that segments are specified by a pair of enum classes (one for process-level segments and one for thread-level segments).