Improve GetCurrentValueAsImpl for integer types, and slightly tweak DetectCallsLookup.
Removed old macro-based specialization of GetCurrentValueAsImpl for integral
types, which required a specialization for each type to be manually specified
(and was causing a bug on Mac OS X, because unsigned long was not specified).

Implemented specializations of GetCurrentValueAsImpl for signed and unsigned
integral types, that use SFINAE to select the appropriate behaviour (based on
signedness).

Modify DetectCallsLookup to explicitly use std::size_t as the backing type for
the Call enum class.