Handle programs that use argc, argv, and envp in main().
Added recording functions to notify SeeC of the values of argc, argv, and envp.
SeeC will make the pointer array and strings for argv readable and writable,
and the pointer array and strings for envp readable, as well as recording the
current memory state in their occupied regions. The memory states are attached
to the FunctionStart event for main(), so they will appear in the ProcessState
when the main() function is entered.