Change the TraceViewerFrame to use a ThreadTime slider in place of a ProcessTime slider when viewing single-threaded traces.
Make SourceViewer show the currently active Instruction when viewing
an updated ThreadState, so that we can see the behaviour of a thread
step-by-step.

Changed lib/Clang/Compile.cpp to compile with debugging enabled.

Added a method to ThreadState to allow moving to a new ThreadTime. This method
is only safe for single-threaded traces currently, as it does not make any
other ThreadState objects participate in the movement.