Change the X-Macro system for runtime errors to include names for the arguments for the runtime error types.
Change the unicode formatting for runtime error descriptions to use named
arguments.

Added new runtime error types for when a bad pointer is passed to a standard
function, so that we can specify that the error is because the user's code is
passing a bad parameter, rather than reading/writing memory itself.