Give all state objects access to their parent objects (e.g. FunctionState can access the parent ThreadState).
Add AllocaState methods to find the associated AllocaInst and the
MemoryState::Region for the alloca'd memory.

Add FunctionState methods to find the llvm::Function.

Change FunctionState::getActiveInstruction() to return a pointer to the active
llvm::Instruction (or a nullptr if no llvm::Instruction is active). The active
index is now found by using getActiveInstructionIndex().

Added FunctionState methods to find the RuntimeValue for a given
llvm::Instruction.

Update the SourceViewerPanel to make use of the new behaviour of
FunctionState::getActiveInstruction().