Show the value of active Instructions as an annotation below the Expr that they belong to.
Added include/seec/Clang/RuntimeValueMapping.hpp, which defines a toString()
function that converts a RuntimeValue into a string representation based on
the Clang type of the Stmt that caused it. Currently only handles a subset of
the builtin types, and pointers.

Added an overloaded SourceViewerPanel::annotateLine() that places an
annotation at a specific line and column. Column positioning isn't natively
supported, so we insert the required number of spaces before each line of
text in the annotation string.