Fix memory state tracing and recreation. State recording now correctly handles states which split existing fragments, and state recreation now correctly handles all situations for forward and reverse iteration. Note that only untyped states are supported still.
The events describing overwritten states have been changed to the following:
StateOverwrite - A complete state was completely replaced.
StateOverwriteFragment - A fragment of a state was completely replaced.
StateOverwriteFragmentTrimmedRight - A state was overwritten on the right-hand side.
StateOverwriteFragmentTrimmedLeft - A state was overwritten on the left-hand side.
StateOverwriteFragmentSplit - A state was overwritten in its interior, causing it to split into two fragments.

seec::trace::MemoryState gained the following functions to facilitate reversing these overwrites:
  unsplit() - Reverse a StateOverwriteFragmentSplit.
  untrimRightSide() - Reverse a StateOverwriteFragmentTrimmedRight.
  untrimLeftSide() - Reverse a StateOverwriteFragmentTrimmedLeft.

Each of the addEvent() overloads for state events in seec::trace::ThreadState is now required to have an additional overload, which takes a second parameter describing the restricted area which the state should be added to.
This overload is used to implement the reversal of StateOverwriteFragment, by adding the overwritten memory state in the specific area of the fragment that was overwritten.
The implementations of removeEvent() for the state overwrite events have been updated to take advantage of this, and now work automatically with all state events that specify the new trait is_memory_state.

The overwritten state information returned by seec::trace::TraceMemoryState has been enriched to support these new events.
Specifically, each overwritten fragment now returns information about the nature of the overwrite: whether it replaced the fragment completely, trimmed the right side, trimmed the left side, or split the fragment.

The implementations of add() in seec::trace::TraceMemoryState and seec::trace::MemoryState have both been simplified to use the clear() methods on these classes to initially clear space for the new state, and then simply add the state.
Some bugs have been fixed in the implementations of clear() on both classes, and they also support the case where a new state splits an existing state fragment.