Start adding memmove state event.
Add the event type for memmove.

Change the execution tracing to use the memmove event.

Make the trace memory state and recreated memory state support memmove events, by the memmove method which moves fragments from a region into a new region and attaches new state information to them.

Can apply the event memmove event when iterating forwards, but not in a specified region.