Finish implementing Memmove state.
Add a new event type, StateCopied, which is subservient to the StateMemmove event and describes a state fragment that was copied from during the event.

Correct the trap implementations of ThreadState::addMemoryState and ThreadState::restoreMemoryState to correctly catch event types that are not memory states.

Add a free function to describe the value of an EventType, seec::trace::describe().

Fix implementations of addMemoryState and restoreMemoryState to attach correct event locations to the created memory states. This requires the methods to accept the EventLocation of the event which they are restoring. The previous implementation would incorrectly attribute all states to the current thread.

Implement addMemoryState and restoreMemoryState for the StateMemmove event type.