Add a new runtime error format select, CStdStream, for describing the IO stream macros available in the C standard library (stdout, stderr, stdin).
Add a new runtime error, UseInvalidStream, to be raised when a call to a standard library function would use a standard stream which is currently closed (e.g. printf() when stdout is closed).

When a ProcessListener is first constructed, register stdout, stderr, and stdin as open streams.

Check that the output streams are valid in the pre-notification for printf() and fprintf().