Make interceptor functions have the same type signature as the intercepted function they replace.
Keep instruction index on the thread environment, so that interceptor functions can find the call site responsible for their call/invocation.

Don't forward pre/post notifications or value updates for calls to interceptor functions (so that the behaviour of indirect calls to interceptor functions matches the behaviour of direct calls).