Added a new runtime error type, NonTerminatedArray, for use when a function expects a NULL-terminated array, but the passed array is not NULL-terminated.
Enabled function interception for execl, execv, execle, execlp, and execvp.
Currently execl, execle, and execlp are not implemented.

Added CStdLibChecker::checkCStringArray(), which takes a pointer and ensures
that a valid array of C strings exists at the destination.

Added parameter error checking and recording for execv, execve, execvp, and
fork.