Add IndexedString methods:
lookupPrimaryIndexAtCharacter(), which gets an iterator to the needle lookup
entry for the primary index at a given character position, or the end of the
needle lookup if no index exists at that position.

getPrimaryIndexAtCharacter() returns the string key of the primary index at a
given character position, or an empty string if no index exists at that
position.