Add ClangEPV explanations of active code to the trace viewer.
Rename the existing indicator "ActiveCode" to "CodeActive".

Add a new indicator type "CodeHighlight", for use when highlighting information
that the user is viewing/interested in.

Add a new control, ExplanationViewer, which inherits from wxStyledTextCtrl, but
should be used by the following (added) methods:
  showExplanation(::clang::Decl const *Decl)
  showExplanation(::clang::Stmt const *Statement)
  clearExplanation()

Move code for setting default settings for common styles, lexer styles, and
indicators into the following functions:
  setupAllSciCommonTypes()
  setupAllSciLexerTypes()
  setupAllSciIndicatorTypes()

Show an ExplanationViewer directly below the source code notebook in the
SourceViewerPanel.