Improve explanation of BinaryOperator expressions.
Add formatAsGeneralTypeString(), which accepts a ::clang::Type const *, and
produces a string describing the general type of the provided type, e.g.
"Integer", "Floating", "Pointer".

Add a new input to BinaryOperator, "type_general", which gives a string
describing the general type of the Expr, generated by the new function
formatAsGeneralTypeString().