Update SimpleWrapper system.
Update value notification to convert integer types to the appropriate
fixed-width type for their size, to avoid ambiguity when calling
TraceThreadListener::notifyValue().

Add a "IgnoreNull" flag to WrappedOutputPointer. If this flag is true, then no
error checking or recording will be performed for NULL values.

Check if the value of errno changes when the original function is called and,
if it does, update the state of errno.