Add llvm::GlobalVariable - clang::ValueDecl mapping information to seec::seec_clang::MappedModule.
This information is already stored by SeeC-Clang, but we weren't reading it
correctly.

Add a new class to represent the mapping, MappedGlobalVariableDecl. This is
analagous to MappedFunctionDecl.

Add a map from llvm::GlobalVariable to MappedGlobalVariableDecl in MappedModule,
and add methods for accessing and querying the map.

Add seec::cm::GlobalVar, which will represent a mapped global variable in a
mapped process state.

Add the cacheClear() method to seec::cm::ProcessState. This should be called
when the underlying state has been moved, to clear cached information about the
mapped interpretation of the state.

Add a ValueStore to seec::cm::ProcessState, which will be used to unique all
seec::cm::Value objects for a given process state. A new store is created when
cacheClear() is called.

When printing a seec::cm::ProcessState, print a list of mapped global variable
names.