Error checking and recording for strtok() and snprintf().
Remove the detect calls stub for strtok, and implement using interception. The
interceptor will raise an error if more than one thread simultaneously calls
strtok, otherwise it will simply error check the inputs, call the original
strtok, and record the state change (the inserted NULL byte - if there is one).

Make the pre-call notification for snprintf() check that the supplied buffer is
writable memory with at least as many bytes available as the caller specified.