Add interception functions for abort, exit, quick_exit, _Exit, atexit, and at_quick_exit.
We keep our own stack of callback functions for atexit and at_quick_exit, and
we manually call them in the intercepted exit and quick_exit.

All functions that terminate the process will first coordinate a thread stop
using the SynchronizedExit, then ensure that the trace is written, before
finally terminating the process.