Start adding the framework for SeeC-Clang-mapped allocas (locals and parameters).
These are (both currently) represented using the new class
seec::cm::AllocaState.

When a seec::cm::FunctionState is created, it will create AllocaState objects
for all of the raw allocas that it can find mapping information for (if it
determines that they are a VarDecl). The FunctionState keeps a list for
parameters (ParmVarDecl mappings) and a list for local variables (all other
mappings). When a FunctionState is printed, it will also show its parameters
and local variables, but at this time the printing of allocas is simply a
placeholder.