Don't show allocas in SeeC-Clang mapped state if they have an associated llvm.dbg.declare that is dominated by the currently active Instruction.
Add a new method to seec::trace::FunctionState, getVisibleAllocas(), which gets
a vector of references to the AllocaState objects for all allocas that are
"visible" (according to the above).

Extend FunctionIndex to store a list of all llvm.dbg.declare Instructions, and
a lookup from Alloca Instructions to the index of associated llvm.dbg.declare
Instructions.