Support using indexed strings for runtime error descriptions.
Remove seec::runtime_errors::format().

New method for getting Unicode descriptions of runtime errors is to use the
class seec::runtime_errors::Description. The Description object holds the
textual description of a RunError, and holds child Description objects for each
of the additional RunError objects attached to the parent RunError.

To simplify the act of combining these textual descriptions (and later, to
support lookup to indices / linked arguments), use the DescriptionPrinterUnicode
class.