Store CompilerInvocation arguments during compilation in seec-clang-test, and use them to create CompilerInvocations when recreating ASTs in MappedModule.
This is a step towards addressing issue #23. We may need to fix up some of the
arguments when recreating (particularly the resources directory). Also, we need
a method to ensure that the include file search always finds the files that
we used in the original compilation (regardless of header search paths on the
system that is doing the recreation).

Remove unused member "CI" from SeeCASTConsumer.

Add seec::seec_clang::getCompileArgumentsDefault(), which handles the generation
of the CompilerInvocation arguments needed to compile a single C99 source file.
This separates the logic from GetCompileForSourceFile (which now defers to this
function).

Add method MappedCompileInfo::getInvocationArguments(), which allows clients to
access the invocation arguments used in the compilation process.

Remove several debugging statements that printed to llvm::errs().

Update MappedModule::getASTForFile() to create a CompilerInvocation from the
arguments stored in the file's MappedCompileInfo.

Modify seec-clang-test to create arguments with getCompileArgumentsDefault(),
create its own CompilerInvocation using these arguments, and finally to save
those arguments into the compile information stored in the generated Module's
metadata.