Proof of concept for asynchronous lazily constructed jQuery.contextMenu menus which request information from the StateGraphViewer using CallbackFSHandler via jQuery's JSONP support.
The current setup is going to have to change because all virtual file systems
are shared in wxWidgets, but our callbacks are going to need to refer to the
local StateGraphViewer object. At the moment the solution appears to be using
a custom protocol for each StateGraphViewer's CallbackFSHandler (probably
"seec" and the StateGraphViewer's runtime address), and instructing the
WebView's webpage to use that specific protocol.